+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.2] - 2019-07-18
 ### Added
+- Added Integer => `bit`
+- Added Integer => `bit?`
+- Added Integer => `bit_clear`
+- Added Integer => `bitmask`
+- Added Integer => `bitmask?`
 - Added Numeric => `close?`
 - Added Numeric => `delimit`
 - Added Numeric => `delta`